+++ lib/layrz_models.dart
@@ -159,6 +159,8 @@ part 'src/trip.dart';
 part 'src/undeclarated_models.dart';
 part 'src/weekdays.dart';
 part 'src/registered_logs.dart';
+part 'src/param_data.dart';
+part 'src/parameter_update.dart';
 
 part 'src/block/block.dart';
 part 'src/block/enums/condition_type.dart';